ENG-5877
Added cast check prior to doing the cast. If not a view, we will send back a small message “not_a_view” which will be aded to the screen_hierarchy during the event reporting. Code below invokes the getParentsOfView() in getParents().

val classJson = JSONObject().put("n", "screenHierarchy")
    .put("v", "${view.getParents()}${NIDServiceTracker.screenName}")

If this is not how we should handle non views during the layout traversal then please comment on how this should be handled (Throw Error or a differnt message perhaps). Thanks.